Update config.py.
Ignore mame_machines.py being in this commit if it shows up, given that I made 0 changes to it. Visual Studio just has it in the changes window. Taunting me. Mocking me for thinking I could improve it and then it turns out I couldn't so  I had to revert. But no revert will ever be good enough in Visual Studio's eyes.