MAME C64: Don't connect a floppy drive if we aren't using one (which we aren't at the moment).
This actually speeds up things quite a bit, especially during loading of tapes, which is exactly when you want to be using turbo to go as fast as possible.

We can leave the tape drive connected if we're not using it, because it doesn't seem to have as much of an overhead, which makes sense if you think about how the drives work and how complex it would be to emulate them.